requirements.txt
anndata==0.10.7
einops==0.8.0
huggingface-hub==0.20.2
joblib==1.3.2
matplotlib==3.7.2
numpy
pandas==2.2.2
Pillow==10.3.0
scanpy==1.10.1
scipy==1.13.0
seaborn==0.13.2
SpatialDM==0.1.0
threadpoolctl==3.5.0
tqdm==4.66.1
timm
# torch==2.0.1+cu118
# torch_geometric==2.3.1
# torchvision==0.15.2+cu118